Move obmd_cfg fixture into test_common.
